% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zh.R
\name{zh}
\alias{zh}
\title{ZH: Fast heatmap plotter as an HTML widget}
\usage{
zh(mat, preserve_order = T, use_viewer = F, display_rownames = T,
  display_colnames = T, pixel_width = 4, pixel_height = 4)
}
\description{
ZH: Fast heatmap plotter as an HTML widget
}

