% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zh.R
\name{zh}
\alias{zh}
\title{ZH: Fast heatmap plotter as an HTML widget}
\usage{
zh(mat, preserve_order = T, use_viewer = F, pixel_width = 5,
  pixel_height = 5)
}
\description{
ZH: Fast heatmap plotter as an HTML widget
}

